+++ src/game/buildings/studio/StudioJobs.h
@@ -20,7 +20,7 @@
 #ifndef WM_STUDIOJOBS_H
 #define WM_STUDIOJOBS_H
 
-#include "jobs/GenericJob.h"
+#include "jobs/BasicJob.h"
 #include "data.h"
 
 struct sFilmPleasureData {
@@ -108,13 +108,9 @@ public:
     sWorkJobResult DoWork(sGirl& girl, bool is_night) override;
     using cBasicJob::cBasicJob;
 protected:
-    Image_Types m_EventImage = IMGTYPE_PROFILE;
+    Image_Types m_EventImage = IMGTYPE_STUDIO_CREW;
 private:
     virtual void HandleUpdate(sGirl& girl, float performance) = 0;
-};
-
-class cMarketResearchJob : public cBasicJob {
-
 };
 
 #endif //WM_STUDIOJOBS_H