+++ src/game/jobs/BasicJob.cpp
@@ -1,0 +1,195 @@
+/*
+ * Copyright 2009, 2010, The Pink Petal Development Team.
+ * The Pink Petal Devloment Team are defined as the game's coders
+ * who meet on http://pinkpetal.org
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include "BasicJob.h"
+#include "xml/getattr.h"
+#include "xml/util.h"
+#include "character/traits/ITraitsManager.h"
+#include "character/sGirl.h"
+#include "cGirls.h"
+#include "CLog.h"
+#include "utils/string.hpp"
+#include "IGame.h"
+
+double cBasicJob::GetPerformance(const sGirl& girl, bool estimate) const {
+    return m_PerformanceData.eval(girl, estimate);
+}
+
+cBasicJob::cBasicJob(JOBS job, const char* xml_file) : IGenericJob(job), m_Interface(this) {
+    if(xml_file) {
+        load_from_xml(xml_file);
+    }
+
+    RegisterVariable("Performance", m_Performance);
+}
+
+void cBasicJob::apply_gains(sGirl& girl, int performance) {
+    m_Gains.apply(girl, performance);
+}
+
+void cBasicJob::load_from_xml(const char* xml_file) {
+    try {
+        load_from_xml_internal(xml_file);
+    } catch (std::exception& error) {
+        g_LogFile.error("job", "Error loading job xml '", xml_file, "': ", error.what());
+        throw;
+    }
+}
+
+void cBasicJob::load_from_xml_internal(const char* xml_file) {
+    DirPath path = DirPath() << "Resources" << "Data" << "Jobs" << xml_file;
+    auto doc = LoadXMLDocument(path.c_str());
+    auto job_data = doc->FirstChildElement("Job");
+    if(!job_data) {
+        throw std::runtime_error("Job xml does not contain <Job> element!");
+    }
+
+    // Info
+    m_Info.ShortName = GetStringAttribute(*job_data, "ShortName");
+    if(const auto* desc_el = job_data->FirstChildElement("Description")) {
+        if(const char* description = desc_el->GetText()) {
+            m_Info.Description = description;
+        } else {
+            g_LogFile.error("jobs", "<Description> element does not contain text. File: ", xml_file);
+        }
+    } else {
+        g_LogFile.error("jobs", "<Job> element does not contain <Description>. File: ", xml_file);
+    }
+
+    std::string prefix = "job." + std::string(xml_file);
+
+    // Performance Criteria
+    const auto* performance_el = job_data->FirstChildElement("Performance");
+    if(performance_el) {
+        m_PerformanceData.load(*performance_el, prefix);
+    }
+
+    // Gains
+    const auto* gains_el = job_data->FirstChildElement("Gains");
+    if(gains_el) {
+        m_Gains.load(*gains_el);
+    }
+
+    // Modifiers
+    const auto* modifiers_el = job_data->FirstChildElement("Modifiers");
+    if(modifiers_el) {
+        // TODO automatically prefix with the jobs name, and allow for loading "local" modifiers
+        // which start with .
+        g_Game->traits().load_modifiers(*modifiers_el, prefix);
+    }
+
+    // Texts
+    const auto* text_el = job_data->FirstChildElement("Messages");
+    if(text_el) {
+        m_TextRepo = ITextRepository::create();
+        m_TextRepo->load(*text_el);
+    }
+    const auto* config_el = job_data->FirstChildElement("Config");
+    if(config_el) {
+        load_from_xml_callback(*config_el);
+    }
+}
+
+const std::string& cBasicJob::get_text(const std::string& prompt) const {
+    assert(m_TextRepo);
+    try {
+        return m_TextRepo->get_text(prompt, m_Interface);
+    } catch (const std::out_of_range& oor) {
+        g_LogFile.error("job", "Trying to get missing text '", prompt, "\' in job ", m_Info.Name);
+        throw;
+    }
+}
+
+bool cBasicJob::has_text(const std::string& prompt) const {
+    return m_TextRepo->has_text(prompt);
+}
+
+std::stringstream& cBasicJob::add_text(const std::string& prompt) {
+    auto& tpl = get_text(prompt);
+    interpolate_string(ss, tpl, [&](const std::string& var) -> std::string {
+        if(var == "name") {
+            return active_girl().FullName();
+        } else if (var == "shift") {
+            return is_night_shift() ? "night" : "day";
+        } else if (m_Replacements.count(var) != 0) {
+            return m_Replacements.at(var);
+        }
+        assert(false);
+    }, rng());
+    return ss;
+}
+
+void cBasicJob::SetSubstitution(std::string key, std::string replace) {
+    m_Replacements[key] = replace;
+}
+
+void cBasicJob::InitWork() {
+    m_Performance = GetPerformance(active_girl(), false);
+}
+
+void cBasicJob::RegisterVariable(std::string name, int& value) {
+    m_Interface.RegisterVariable(std::move(name), value);
+}
+
+bool cBasicJobTextInterface::LookupBoolean(const std::string& name) const {
+    return m_Job->active_girl().has_active_trait(name.c_str());
+}
+
+int cBasicJobTextInterface::LookupNumber(const std::string& name) const {
+    auto split_point = name.find(':');
+    auto type = name.substr(0, split_point);
+    if(type == "stat") {
+        return m_Job->active_girl().get_stat(get_stat_id(name.substr(split_point+1)));
+    } else if(type == "skill") {
+        return m_Job->active_girl().get_skill(get_skill_id(name.substr(split_point+1)));
+    } else if (type.size() == name.size()) {
+        try {
+            return *m_MappedValues.at(name);
+        } catch (const std::out_of_range& oor) {
+            g_LogFile.error("job", "Unknown job variable '", name, '\'');
+            BOOST_THROW_EXCEPTION(std::runtime_error("Unknown job variable: " + name));
+        }
+    } else {
+        g_LogFile.error("job", "Unknown value category ", type, " of variable ", name);
+        BOOST_THROW_EXCEPTION(std::runtime_error("Unknown value category: " + type));
+    }
+}
+
+void cBasicJobTextInterface::SetVariable(const std::string& name, int value) const {
+    int* looked_up = m_MappedValues.at(name);
+    *looked_up = value;
+}
+
+void cBasicJobTextInterface::TriggerEvent(const std::string& name) const {
+    throw std::logic_error("Event triggers are not implemented yet");
+}
+
+void cBasicJobTextInterface::RegisterVariable(std::string name, int& value) {
+    m_MappedValues[std::move(name)] = &value;
+}
+
+IGenericJob::eCheckWorkResult cBasicJob::SimpleRefusalCheck(sGirl& girl, Action_Types action) {
+    if (girl.disobey_check(action, job()))
+    {
+        add_text("refuse");
+        girl.AddMessage(ss.str(), IMGTYPE_PROFILE, EVENT_NOWORK);
+        return eCheckWorkResult::REFUSES;
+    }
+    return eCheckWorkResult::ACCEPTS;
+}