+++ src/game/jobs/Cleaning.cpp
@@ -17,7 +17,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "GenericJob.h"
+#include "BasicJob.h"
 #include "character/sGirl.h"
 #include "buildings/IBuilding.h"
 #include "cGirls.h"