+++ src/game/jobs/GenericJob.h
@@ -20,15 +20,11 @@
 #ifndef WM_GENERICJOB_H
 #define WM_GENERICJOB_H
 
-#include <utility>
 #include <vector>
 #include <memory>
 #include <sstream>
-#include <unordered_map>
 #include "cJobManager.h"
 #include "Constants.h"
-#include "text/repo.h"
-#include "JobData.h"
 #include <boost/variant.hpp>
 
 
@@ -91,8 +87,28 @@ protected:
         IMPOSSIBLE
     };
 
+    sGirl& active_girl();
     const sGirl& active_girl() const;
     bool is_night_shift() const;
+
+    // resources
+    //  bulk resources
+    /// Consumes up to `amount` points of the given resource. If less is available,
+    /// that amount will we consumed. Returns the amount of actual consumption.
+    int ConsumeResource(const std::string& name, int amount);
+
+    /// Provides `amount` points of the given resource.
+    void ProvideResource(const std::string& name, int amount);
+
+    /// Tries to consume `amount` of the given resource. If not enough is available,
+    /// no resource is consumed and false is returned.
+    bool TryConsumeResource(const std::string& name, int amount);
+
+    //  one-on-one interactions
+    void ProvideInteraction(const std::string& name, int amount);
+    sGirl* RequestInteraction(const std::string& name);
+
+    bool HasInteraction(const std::string& name) const;
 
 private:
     virtual void InitWork() {}
@@ -112,64 +128,6 @@ private:
 
 protected:
     sJobInfo m_Info;
-};
-
-class cBasicJob;
-
-class cBasicJobTextInterface : public IInteractionInterface {
-public:
-    cBasicJobTextInterface() = delete;
-    explicit cBasicJobTextInterface(cBasicJob* job) : m_Job(job) {}
-
-    bool LookupBoolean(const std::string& name) const final;
-
-    int LookupNumber(const std::string& name) const final;
-
-    void TriggerEvent(const std::string& name) const final;
-    void SetVariable(const std::string& name, int value) const final;
-
-    void RegisterVariable(std::string name, int& value);
-private:
-    std::unordered_map<std::string, int*> m_MappedValues;
-    cBasicJob* m_Job;
-};
-
-class cBasicJob : public IGenericJob {
-public:
-    cBasicJob(JOBS job, const char* xml_file = nullptr);
-    double GetPerformance(const sGirl& girl, bool estimate) const override;
-
-protected:
-    void apply_gains(sGirl& girl, int performance);
-
-    void load_from_xml(const char* xml_file);
-    const std::string& get_text(const std::string& prompt) const;
-    bool has_text(const std::string& prompt) const;
-    std::stringstream& add_text(const std::string& prompt);
-
-    // processing variables
-    void InitWork() override;
-    void RegisterVariable(std::string name, int& value);
-
-    void SetSubstitution(std::string key, std::string replace);
-
-    int m_Performance;
-
-    eCheckWorkResult SimpleRefusalCheck(sGirl& girl, Action_Types action);
-
-private:
-    cJobPerformance m_PerformanceData;
-    cJobGains       m_Gains;
-
-    std::unique_ptr<ITextRepository> m_TextRepo;
-    cBasicJobTextInterface m_Interface;
-
-    void load_from_xml_internal(const char* xml_file);
-    virtual void load_from_xml_callback(const tinyxml2::XMLElement& job_element) {};
-
-    friend class cBasicJobTextInterface;
-
-    std::unordered_map<std::string, std::string> m_Replacements;
 };
 
 #endif //WM_GENERICJOB_H